# src/src-vpn/docs/api.yaml
openapi: "3.0.3"
info:
  title: Unkillable Messenger API
  version: "1.0.0"
  description: REST API for the Unkillable Messenger platform.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/peers:
    get:
      summary: List WireGuard peers
      operationId: listPeers
      responses:
        "200":
          description: List of peers
    post:
      summary: Add a peer
      operationId: addPeer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                publicKey: { type: string }
                endpoint: { type: string }
...
  /api/files/upload:
    post:
      summary: Upload a file
      operationId: uploadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: File uploaded
...
  /api/stream/create:
    post:
      summary: Create a stream
      operationId: createStream
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                channelName: { type: string }
      responses:
        "200":
          description: Stream created
...
  /api/stream/end:
    post:
      summary: End a stream
      operationId: endStream
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                streamId: { type: string }
      responses:
        "200":
          description: Stream ended
...